test(otel): share collector fixture and preserve routing cases
