+++ extensions/diagnostics-otel/src/service.otlp-export.test.ts
@@ -9,8 +9,6 @@
 // Collector-boundary cases run owned mode, which now composes private providers and never
 // registers global SDK state; teardown still restores the preloaded globals for trace cases.
 import { mkdtempSync, rmSync, writeFileSync } from "node:fs";
-import { createServer, type IncomingHttpHeaders } from "node:http";
-import type { AddressInfo } from "node:net";
 import { tmpdir } from "node:os";
 import path from "node:path";
 import { context, diag, DiagLogLevel, metrics, propagation, trace } from "@opentelemetry/api";
@@ -40,6 +38,7 @@ import {
   createOtelContext,
   emitRealSdkSignals,
   startOtelService,
+  startOtlpReceiver,
   stopStartedOtelServices,
 } from "./service.test-helpers.js";
 
@@ -199,43 +198,6 @@ function captureOtelDiagnostics(): string[] {
   return messages;
 }
 
-async function startOtlpReceiver() {
-  const requests: Array<{
-    contentType: string | undefined;
-    headers: IncomingHttpHeaders;
-    method: string | undefined;
-    url: string;
-  }> = [];
-  const server = createServer((request, response) => {
-    request.resume();
-    request.on("end", () => {
-      requests.push({
-        contentType: request.headers["content-type"],
-        headers: request.headers,
-        method: request.method,
-        url: request.url ?? "",
-      });
-      response.writeHead(200, { "content-type": "application/x-protobuf" });
-      response.end();
-    });
-  });
-  await new Promise<void>((resolve, reject) => {
-    server.once("error", reject);
-    server.listen(0, "127.0.0.1", resolve);
-  });
-  const { port } = server.address() as AddressInfo;
-  return {
-    endpoint: `http://127.0.0.1:${port}`,
-    requests,
-    async close() {
-      await new Promise<void>((resolve, reject) => {
-        server.close((error) => (error ? reject(error) : resolve()));
-        server.closeIdleConnections();
-      });
-    },
-  };
-}
-
 function releasePreloadedOtelGlobals() {
   context.disable();
   logs.disable();
@@ -305,7 +267,7 @@ const SHARED_ENDPOINT_ROUTING_CASES = [
 
 test.each(
   SHARED_ENDPOINT_ROUTING_CASES.flatMap((entry) =>
-    (["config", "environment"] as const).map((source) => ({ ...entry, source })),
+    (["config", "environment"] as const).map((source) => Object.assign({ source }, entry)),
   ),
 )(
   "routes real exporters from a shared $label endpoint in $source",
